requirements.txt
python-telegram-bot==21.1
google-generativeai
supabase
flask
python-dotenv
httpx[http2]